Refactor service and result selectors to utilize parameterized ordering selector
Utilize customPosition in distance ordering
