+++ src/redux/selectors/service.js
@@ -1,5 +1,6 @@
 import { createSelector } from 'reselect';
 import { filterCities } from '../../utils/filters';
+import getOrderedData from './ordering';
 
 
 const units = state => state.service.data;
@@ -9,13 +10,17 @@ const cities = state => [
   ...state.settings.espoo ? ['espoo'] : [],
   ...state.settings.kauniainen ? ['kauniainen'] : [],
 ];
+const userLocation = state => state.user.position.coordinates;
+const customLocation = state => state.user.customPosition.coordinates;
 
-export const getServiceUnits = createSelector(
-  [units, cities],
-  (units, cities) => {
+export const getServiceUnits = state => createSelector(
+  [units, cities, userLocation, customLocation],
+  (units, cities, userLocation, customLocation) => {
     const filteredUnits = units.filter(filterCities(cities, true));
-    return filteredUnits;
+    const location = customLocation || userLocation || null;
+    const orderedUnits = getOrderedData(filteredUnits, location)(state);
+    return orderedUnits;
   },
-);
+)(state);
 
 export default { getServiceUnits };
